requirements.txt
# Core dependencies
python-dotenv==1.0.0

# LLM providers
google-generativeai==0.3.2
ollama==0.1.6

# Embeddings
sentence-transformers==2.7.0

# Vector store
faiss-cpu==1.9.0
numpy>=1.26.0

# Database
sqlalchemy==2.0.25

# Utilities
colorama==0.4.6

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
